+++ CHANGELOG.md
@@ -31,6 +31,7 @@
 - Authentication support for Kafka connection. ([#530](https://github.com/getsentry/vroom/pull/530))
 - Add support for android chunks. ([#540](https://github.com/getsentry/vroom/pull/540))
 - Use generic Chunk interface in CallTreesReadJob ([#554](https://github.com/getsentry/vroom/pull/554))
+- Add start/end to profile example. ([#575](https://github.com/getsentry/vroom/pull/575))
 
 **Bug Fixes**: